// src/app/dashboard/users/interfaces/users.interfaces.ts
export interface IUser {
  userId?: number;
  dni: string;
  firstname: string;
  lastname: string;
  username: string;
  role: string;
  password: string;
}

// Generated by https://quicktype.io

export interface IUserResponse {
  userId:                number;
  dni:                   string;
  firstname:             string;
  lastname:              string;
  role:                  null;
  username:              string;
  password:              string;
  createdAt:             Date;
  enabled:               boolean;
  authorities:           any[];
  accountNonLocked:      boolean;
  credentialsNonExpired: boolean;
  accountNonExpired:     boolean;
}